#130 Extract battle timer conversion into timer.rs
